Moved each parser to its own file
