+++ src/parse/symbol/expression/assignment.rs
@@ -1,0 +1,35 @@
+//! Assignment parser
+
+use lex::{tokens, Token, Tokenizer, TokenType, TokenData};
+use parse::{Parser, ParseResult, ParseError, Precedence};
+use parse::ast::*;
+use parse::symbol::InfixParser;
+
+/// Parses a declaration/// Parses an assignment expresion.
+///
+/// # Examples
+/// ```text
+///   x    =   y + 2
+/// (left) ^ ->right:expression
+/// ```
+#[derive(Debug)]
+pub struct AssignmentParser { }
+impl<T: Tokenizer> InfixParser<Expression, T> for AssignmentParser {
+    fn parse(&self, parser: &mut Parser<T>,
+             left: Expression, _token: Token) -> ParseResult<Expression> {
+        debug_assert!(_token.text == tokens::Equals,
+                      "Assign parser called with non-assign token {:?}", _token);
+        let ident = try!(left.expect_identifier());
+        let right_expr = try!(parser.expression(Precedence::Assign));
+        let right = try!(right_expr.expect_value());
+        Ok(Expression::Assignment(Assignment::new(ident, Box::new(right))))
+    }
+    fn get_precedence(&self) -> Precedence {
+        Precedence::Assign
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+}
