+++ src/parse/symbol/expression/literal.rs
@@ -1,0 +1,53 @@
+//! Literal expression parsing
+
+use lex::{Token, Tokenizer, TokenData, TokenType};
+use parse::{Parser, ParseError, ParseResult};
+use parse::symbol::PrefixParser;
+use parse::ast::*;
+
+/// Returns a literal expression
+///
+/// # Examples
+/// ```text
+/// 34
+/// ^literal
+/// ```
+pub struct LiteralParser { }
+impl<T: Tokenizer> PrefixParser<Expression, T> for LiteralParser {
+    fn parse(&self, _parser: &mut Parser<T>, token: Token) -> ParseResult<Expression> {
+        match token.data {
+            TokenData::NumberLiteral(_) =>
+                Ok(Expression::Literal(Literal::new(token))),
+            _ => Err(ParseError::ExpectedToken {
+                    expected: TokenType::Literal,
+                    got: token
+                })
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use lex::{Token, TokenData, TokenType};
+    use parse::ast::{Expression, Literal};
+    use parse::symbol::LiteralParser;
+    use parse::tests as parse_tests;
+
+    #[cfg(test)]
+    fn it_parses_literal_number() {
+        let mut parser = parse_tests::parser("5");
+        let expected_token = Token {
+            data: TokenData::NumberLiteral(5f64),
+            .. Default::default()
+        };
+        let expected = Expression::Literal(Literal::new(expected_token.clone()));
+        let parsed = LiteralParser { }.parse(&mut parser, expected_token).unwrap();
+        parse_tests::ensure_expression_eq(expected, parsed);
+    }
+
+    #[cfg(test)]
+    fn it_fails_non_number_token() {
+        let mut parser = parse_tests::eof_parser();
+        // TODO test
+    }
+}
