+++ src/parse/symbol/expression/mod.rs
@@ -1,0 +1,18 @@
+mod literal;
+mod identifier;
+mod parens;
+mod assignment;
+mod assign_op;
+mod declaration;
+
+pub use self::literal::LiteralParser;
+pub use self::identifier::IdentifierParser;
+pub use self::parens::ParensParser;
+pub use self::assignment::AssignmentParser;
+pub use self::assign_op::AssignOpParser;
+pub use self::declaration::DeclarationParser;
+
+#[cfg(test)]
+mod tests {
+
+}
